+++ NEWS.md
@@ -1,3 +1,11 @@
+Nov 18th, 2020
+==============
+* Version `0.39.2` of the server extension
+* Version `0.23.4` of the Python library
+    * Better messages on remote errors.
+    * Fix error on unknown status codes.
+    * Here bulk geocoder: Return also errors and no-matches.
+
 Sep 23th, 2020
 ==============
 * Version `0.39.1` of the server extension